--- a/executor/aggregate.go
+++ b/executor/aggregate.go
@@ -31,6 +31,7 @@ import (
 	"github.com/pingcap/tidb/types/json"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/codec"
+	"github.com/pingcap/tidb/util/collate"
 	"github.com/pingcap/tidb/util/logutil"
 	"github.com/pingcap/tidb/util/memory"
 	"github.com/pingcap/tidb/util/set"
@@ -1258,16 +1259,21 @@ func (e *vecGroupChecker) evalGroupItemsAndResolveGroups(item expression.Express
 			lastRowDatum.SetMysqlJSON(col.GetJSON(numRows - 1).Copy())
 		}
 	case types.ETString:
-		previousKey := codec.ConvertByCollationStr(col.GetString(0), tp)
+		ctor := collate.GetCollator(tp.Collate)
+		buf := &collate.Buffer{}
+		buf2 := &collate.Buffer{}
+		previousKey := ctor.KeyByBytes(buf, col.GetBytes(0))
 		for i := 1; i < numRows; i++ {
-			key := codec.ConvertByCollationStr(col.GetString(i), tp)
+			buf2.Reset()
+			key := ctor.KeyByBytes(buf2, col.GetBytes(i))
 			isNull := col.IsNull(i)
 			if e.sameGroup[i] {
-				if isNull != previousIsNull || previousKey != key {
+				if isNull != previousIsNull || bytes.Compare(previousKey, key) != 0 {
 					e.sameGroup[i] = false
 				}
 			}
-			previousKey = key
+			buf.Reset()
+			previousKey = buf.SetByte(key)
 			previousIsNull = isNull
 		}
 		if !firstRowIsNull {