<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
Close #17949
Optimize the performance of CI collation stuff.


### What is changed and how it works?

1. In evalGroupItemsAndResolveGroups, use `Comapre()` instead of `ConvertByCollationStr()`
. The internal test shows that it's faster.
2. Rewriter `Compare` function, including `mapTable`.
3. Give a memory friendly function `KeyByBytes` to avoid memory allocation, it will be used in the future.


Here is the benchmark:
```
BenchmarkBinKeyByBytesPadding-12      	11154472	        91.3 ns/op
BenchmarkBinKeyByBytesNoPadding-12    	26436696	        45.0 ns/op
BenchmarkCIKeyByBytes-12              	 1548896	       755 ns/op
BenchmarkMasterCIKey-12               	  452618	      2440 ns/op
BenchmarkBinKeyNoPadding-12           	 4768342	       248 ns/op
```


- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

No test

Side effects



### Release note <!-- bugfixes or new feature need a release note -->

- optimize the performance of CI collation stuff