@zz-jason 2021-02-09
I'm going to close this PR since it hasn't been updated for a long time. feel free to reopen if you want to continue with it. thank you for your contribution.

BTW, I think this is a performance optimization instead of a bugfix, it's no need to be cherry-picked to release branches.